+++ components/webui/imports/ui/SearchView/SearchResults/index.jsx
@@ -2,8 +2,8 @@ import {useCallback} from "react";
 import Container from "react-bootstrap/Container";
 import Row from "react-bootstrap/Row";
 
-import SearchResultsHeader from "./SearchResultsTable/SearchResultsHeader";
-import SearchResultsTable from "./SearchResultsTable/SearchResultsTable";
+import SearchResultsHeader from "./SearchResultsHeader";
+import SearchResultsTable from "./SearchResultsTable";
 import SearchResultsTimeline from "./SearchResultsTimeline";
